Refactor: Seperate the Attack and move logic
