+++ logic/GameEngine/Map.cs
@@ -20,6 +20,15 @@ namespace GameEngine
 			Color3 = 3,
 			Color4 = 4,
 		}
+		public static ColorType TeamToColor(long teamID)
+		{
+			return (ColorType)(teamID + 1L);
+		}
+		public static long ColorToTeam(ColorType color)
+		{
+			return (long)color - 1L;
+		}
+
 
 		private ColorType[,] cellColor;         //储存每格的颜色
 		public ColorType[,] CellColor
