+++ logic/Gaming/Game.MoveManager.cs
@@ -1,0 +1,68 @@
+﻿using GameEngine;
+using System;
+using System.Collections;
+using System.Threading;
+using THUnity2D;
+
+namespace Gaming
+{
+	public partial class Game
+	{
+		private MoveManager moveManager;
+		private class MoveManager
+		{
+
+			//人物移动
+			public void MovePlayer(Character playerToMove, int moveTimeInMilliseconds, double moveDirection)
+			{
+				moveEngine.MoveObj(playerToMove, moveTimeInMilliseconds, moveDirection);
+			}
+
+
+			private void ActivateMine(Character player, Mine mine)
+			{
+				gameMap.ObjListLock.EnterWriteLock();
+				try { gameMap.ObjList.Remove(mine); }
+				catch { }
+				finally { gameMap.ObjListLock.ExitWriteLock(); }
+
+				switch (mine.GetPropType())
+				{
+					case PropType.Dirt:
+						player.AddMoveSpeed(Map.Constant.dirtMoveSpeedDebuff, Map.Constant.buffPropTime);
+						break;
+					case PropType.Attenuator:
+						player.AddAP(Map.Constant.attenuatorAtkDebuff, Map.Constant.buffPropTime);
+						break;
+					case PropType.Divider:
+						player.ChangeCD(Map.Constant.dividerCdDiscount, Map.Constant.buffPropTime);
+						break;
+				}
+			}
+
+
+			private Map gameMap;
+			private MoveEngine moveEngine;
+			public MoveManager(Map gameMap)
+			{
+				this.gameMap = gameMap;
+				this.moveEngine = new MoveEngine
+				(
+					gameMap: gameMap,
+					OnCollision: (obj, collisionObj, moveVec) =>
+					{
+						if (collisionObj is Mine)
+						{
+							ActivateMine((Character)obj, (Mine)collisionObj);
+						}
+						return false;
+					},
+					EndMove: obj =>
+					{
+						GameObject.Debug(obj, " end move at " + obj.Position.ToString() + " At time: " + Environment.TickCount64);
+					}
+				);
+			}
+		}
+	}
+}
